Do not send reaction error messages to DMs
This method of error reporting is unreliable as it does not verify that the user can receive DMs, and if they can't this will throw an error of its own that is entirely unhandled and causes a crash. It is expected that reaction handlers do not use thrown errors to control user flow and so the only thrown errors are internal errors to be logged. This method is a bit slap-dash but can be improved when the wrapper system is overhauled.