Refactor: Discord events as modules, role provider modules, commands receive Eris Messages (#147)
* Experimental events architecture: move code to events and index
* Remove logging from TournamentManager.authenticateX; all commands logged immediately after for the command attempt
* Pass Eris.Message to commands, add reply util
* getMentionedUser -> firstMentionOrFail
* Add TournamentManager.getConfirmed, move dump, pie, players logic into commands
* Add OrganiserRoleProvider
* AddParticipantRoleProvider
* Fix/suppress tests

We adopt a per-file approach to events. Unfortunately this means that we'll have to add the typings and make sure they match Eris. Currently passing in dependencies via a higher-order function as a sort of dependency injection but this is very much changeable as we figure out the best approach.

All custom typings that messageCreate uses have been pulled over in an effort to isolate; it remains to be seen whether copying them will be fruitful. They will eventually be removed. The skipped tests will also be replaced by command unit tests later.

Role providers are based on previous work in the 46 branch.