Move getConfirmed to Database
Arguably this is a needless indirection for a simple task and could be removed entirely. However, keeping a function like this makes testing easier because we have something to stub that has a smaller return signature, so we don't have to mock up a whole DatabaseTournament. Also, since its in the Database file now, it can use a dedicated query, though I'm not confident that I've done it correctly.