Extract simple commands from TournamentManager (#192)
* Extract bye operations from TournamentManager
* Extract channel operations
* Extract authentication operations
* Extract tournament list
* Extract updateTournament
* Extract host operations
* Extract synchronise

* Move getConfirmed to Database
Arguably this is a needless indirection for a simple task and could be removed entirely. However, keeping a function like this makes testing easier because we have something to stub that has a smaller return signature, so we don't have to mock up a whole DatabaseTournament. Also, since it's in the Database file now, it can use a dedicated query.