Create DeckManager class for CommandSupport (#194)
* Create DeckManager class for CommandSupport

This allows proper stubbing and spying of Deck-related functions and the extraction of further commands from TournamentManager. This also consolidates deck and discordDeck into one file. This is a bit of a monolithic commit because the way everything was intertwined made it hard to be atomic, sorry. The test for the deck command has been expanded.

* Move splitText function

deck.ts is the only file using it, so for now there's no need to have it in the general utility file. We can export it again if we need to in the future, however.