Add command mc!info for in-server use by all (#314)
This is a comprehensive way to access details like the Challonge link, participant limit, number of registrations, registered byes, and who's in charge. It's allowed for everybody because the id has to be known in the first place. Cannot be used in direct messages to figure out all tournaments. Will not ping any mentioned user. createTournamentEmbed is a reusable pure function for the future.