Fix all attachments broken by switching to Discord.js (#327)
String type for BufferResolvable is treated as file name, not contents: https://discord.js.org/#/docs/main/stable/typedef/BufferResolvable

This resulted in the program attempting to find a file in its filesystem with a long name equal to the contents of the deck, which always failed due to the long names. The errors often failed to be sent to the Discord webhook, since the payload contained the entire deck string, exceeding the size limit.